// src/types/playerAppearance.ts
export interface PlayerAppearance {
	key_id: string;
	tournament_id: string;
	tournament_name: string;
	match_id: string;
	match_name: string;
	match_date: string;
	stage_name: string;
	group_name: string;
	team_id: string;
	team_name: string;
	team_code: string;
	home_team: string;
	away_team: string;
	player_id: string;
	family_name: string;
	given_name: string;
	shirt_number: string;
	position_name: string;
	position_code: string;
	starter: string;
	substitute: string;
}